requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings
email-validator
passlib[bcrypt]
jose
